requirements.txt
Flask
SQLAlchemy
requests
PyJWT
Werkzeug
psycopg2-binary
